[Adds abstract ObjectSpawner] & [Makes JackolanternSpawner a subclass of ObjectSpawner]
