+++ Assets/Resources/Spawnables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6e900d1b685d6df4aa62d0eed6ed732a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
