+++ Assets/Resources/Spawnables/Enemies/JackolanternSpawningData.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e6ee92466e830624ca741a992f6ef849, type: 3}
-  m_Name: JackolanternData
+  m_Script: {fileID: 11500000, guid: 266d6e5731ed3c44a882951e86f18ef5, type: 3}
+  m_Name: JackolanternSpawningData
   m_EditorClassIdentifier: 
   minimumRespawnTime: 1.8
   maximumRespawnTime: 3.3
