+++ Assets/Resources/Spawnables/Enemies/JackolanternSpawningData.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e2e1347a5afc52d4abd61810a306ec9d
+guid: 2f2892da7c09db64f9b47d40f65c778b
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
