+++ Assets/Resources/Spawnables/Obstacles.meta
