+++ Assets/Resources/Spawnables/Obstacles/SmallObstaclesSpawningData.asset
