+++ Assets/Scripts/Shared/ScriptableObjectsDefinitions/SpawningDataScriptable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 266d6e5731ed3c44a882951e86f18ef5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
